bugfix: script to build docker image was deleting up-to-date image
script to build and replace capstone-runner docker image was also
deleting old images.
the image ID for two identical images are the same. even if they it
is built fresh. this case wasn't considered in the script, and it
would delete the image with old ID.
now it compares whether this old ID is different from new ID before
deleting.