frontend-refactor: fix extra </div> tag, put raw html inside html.HTML
By default, strings must be interpreted as html.Text in Kutty. This
must be escaped. It is currently not escaped and the code runs fine,
but the expected type of raw html must be html.HTML.

This commit also fixes the indentation for the raw HTML block for
hero. There was an extra </div> closing tag, which closed the "page"
div. This was accidental and has been reverted in this commit.