Add fixtures for FakeRedis: clear_queue and worker
clear_queue is an autouse fixture that will empty the queue before
each new test is run.
worker is a fixture for getting a handle on the simple worker.
this won't run the tasks by itself, but the tasks should instead be
run with `worker.work(burst=True)` when needed. this is to run the
worker without forking, and to allow tasks to finish before
proceeding with the next checks.