#57: remove time.sleep from tests
There was a time.sleep() each time a project was updated from git
in the tests. This was to give some time to capstone webapp to
get the webhook call from Gitto and make these changes. But this is
unnecessary because tasks don't run in async mode in test environment.
The task runs in the webhook call from Gitto, and delays the completion
of the "git push" command.